+++ client/src/features/relatedLists/FacetsRelatedList.tsx
@@ -1,9 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import React, { useState } from 'react'
+import React from 'react'
 
 import config from '../../config/config'
 import DescriptionList from '../../styles/shared/DescriptionList'
-import { IOrderedItems, ISearchResults } from '../../types/ISearchResults'
 import {
   getCriteriaFromHalLink,
   getResultTabFromHalLink,
@@ -14,8 +13,16 @@ import ListItem from './ListItem'
 interface IProps {
   url: string
   searchTerm: string
-  data: ISearchResults
+  data: {
+    requests: {
+      [key: string]: Array<{ value: string; totalItems: number }>
+    }
+    total: number
+  }
   title: string
+  page: number
+  lastPage: number
+  setPage: (x: number) => void
 }
 
 const FacetsRelatedList: React.FC<IProps> = ({
@@ -23,14 +30,16 @@ const FacetsRelatedList: React.FC<IProps> = ({
   searchTerm,
   data,
   title,
+  page,
+  lastPage,
+  setPage,
 }) => {
   const criteria = getCriteriaFromHalLink(url, 'facets')
   const scope = getResultTabFromHalLink(url)
 
-  const unitLength = 20
-  const [displayLength, setDisplayLength] = useState(unitLength)
-
-  const recordLinks = (facets: Array<IOrderedItems>): any => {
+  const recordLinks = (
+    facets: Array<{ value: string; totalItems: number }>,
+  ): any => {
     const filteredFacets = facets.filter((facet) => {
       let { value } = facet
       // type cast as a string
@@ -49,13 +58,12 @@ const FacetsRelatedList: React.FC<IProps> = ({
     }
     return (
       <React.Fragment>
-        {filteredFacets
-          .slice(0, displayLength)
-          .map((facet: IOrderedItems, ind: number) => {
+        {filteredFacets.map(
+          (facet: { value: string; totalItems: number }, ind: number) => {
             const { value, totalItems } = facet
             return (
               <ListItem
-                key={facet.id}
+                key={value}
                 uri={value as string}
                 count={totalItems || 0}
                 title={title}
@@ -66,21 +74,22 @@ const FacetsRelatedList: React.FC<IProps> = ({
                 itemSpacing="double"
               />
             )
-          })}
-        {displayLength >= unitLength && displayLength < facets.length && (
+          },
+        )}
+        {page !== lastPage && (
           <button
             type="button"
             className="btn btn-link show-more"
-            onClick={() => setDisplayLength(facets.length)}
+            onClick={() => setPage(page + 1)}
           >
-            Show All
+            Show More
           </button>
         )}
-        {displayLength > 20 && (
+        {page !== 1 && (
           <button
             type="button"
             className="btn btn-link show-less"
-            onClick={() => setDisplayLength(unitLength)}
+            onClick={() => setPage(page - 1)}
           >
             Show Less
           </button>
@@ -89,18 +98,20 @@ const FacetsRelatedList: React.FC<IProps> = ({
     )
   }
 
-  const { orderedItems } = data
-  if (orderedItems !== null && orderedItems.length > 0) {
-    return (
-      <DescriptionList
-        data-testid={`related-facets-description-list-${searchTerm}`}
-      >
-        {recordLinks(orderedItems)}
-      </DescriptionList>
-    )
-  }
+  const facetsData: Array<{ value: string; totalItems: number }> = []
+  Object.values(data.requests).map((value) =>
+    value.map((v) =>
+      facetsData.push({ value: v.value, totalItems: v.totalItems }),
+    ),
+  )
 
-  return null
+  return (
+    <DescriptionList
+      data-testid={`related-facets-description-list-${searchTerm}`}
+    >
+      {recordLinks(facetsData)}
+    </DescriptionList>
+  )
 }
 
 export default FacetsRelatedList